openrouter/video-automation/requirements.txt
pytrends==5.13.0
yt-dlp==2024.8.23
ffmpeg-python==0.2.0
librosa==0.10.2
faster-whisper==1.0.1
edge-tts==1.0.9
python-dotenv==1.0.0
google-auth==2.34.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.135.0
moviepy==0.14.0
pillow==10.3.0
APScheduler==3.10.4
python-json-logger==2.0.7